+++ src/addons/display_commander/ui/new_ui/hotkeys_tab.hpp
@@ -27,8 +27,6 @@ enum class HotkeyId : int {
     DisplayCommanderUi,
     PerformanceOverlay,
     Stopwatch,
-    SystemVolumeUp,
-    SystemVolumeDown,
     WinDown,
     WinUp,
     WinLeft,